Fix exception loading .NET Portable profiles on Mac.
In Mono 3.2.3 on the Mac there are two profiles with the same supported
frameworks:

    Profile7 and ProfileXI7
    net45+win+MonoAndroid10+MonoTouch10

NuGet adds these into a dictionary using the supported frameworks as a key
and expects the target frameworks to be different for all profiles. The
dictionary lookup is not required for this utility application so it has
been removed.